fixed according to comments
