+++ src/main/java/com/example/Feline.java
@@ -15,7 +15,7 @@ public class Feline extends Animal implements Predator {
     }
 
     public int getKittens() {
-        return getKittens(1);
+        return getKittens(4);
     }
 
     public int getKittens(int kittensCount) {
